Adding $switch and rearranging parser somewhat. Removing the need for "store" of context.
